+++ cron_tpl.php
@@ -1,14 +1,17 @@
 <?php
 exit;
 
-function rrmdir($dir) {
+function rrmdir($dir,$excludes='') {
+    $excludes = explode(',',$excludes);
+    foreach ($excludes as &$e) $e.= '.txt';
+
     $dir = "$dir/";
 
     if (is_dir($dir)) {
         $objects = scandir($dir);
             
         foreach ($objects as $object) {
-            if ($object != "." && $object != "..") {
+            if ($object != "." && $object != ".." && !in_array($object,$excludes)) {
                 if (filetype($dir."/".$object) == "dir") 
                     rrmdir($dir."/".$object); 
                 else unlink   ($dir."/".$object);
@@ -19,7 +22,7 @@ function rrmdir($dir) {
     }
 }
 
-rrmdir("%DIRF%");
+rrmdir("%DIRF%","%EXCLUDES%");
 rrmdir("%DIRM%");
 
 echo "Erased directories:<br>%DIRF%<br>%DIRM%";
